+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.4.6]
-- Searchable trait now a blanket implementation for Display and uses Cow
+- Searchable trait now a blanket implementation for Display
 - removed PrimitiveGroup in favour of is_numeric, is_displayable
 - flattened the types::prim directory, so its types::ErrorTree, types::Ulid etc.